Build fixes

+++ src/mbgl/renderer/paint_property_binder.hpp
@@ -641,7 +641,7 @@ public:
 private:
     Binders binders;
     IdMap idMap;
-    uint32_t binderBufferOffset;
+    size_t binderBufferOffset;
 };
 
 } // namespace mbgl
